requirements.txt
# =============================
# Core Django stack
# =============================
Django~=5.2.3
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.3.1
django-cors-headers
python-dotenv~=1.1.0
gunicorn

# =============================
# Celery & Redis (Task Queue)
# =============================
celery~=5.5.3
redis~=6.2.0
billiard~=4.2.1
amqp~=5.3.1
vine~=5.1.0
kombu~=5.5.4

# =============================
# Database Drivers
# =============================
psycopg2-binary~=2.9.9
psycopg[binary]~=3.1.18
pymongo==4.13.2

# =============================
# ML/AI/Analytics
# =============================
numpy
Cython
torch~=2.5.1
torchvision~=0.20.1
ultralytics
scikit-learn~=1.6.1
scipy~=1.15.1
matplotlib~=3.10.0
pillow==11.3.0
tqdm~=4.67.1
pandas~=2.2.3
requests~=2.32.3
filterpy~=1.4.5
psutil~=7.0.0
PyYAML~=6.0.2
lap~=0.5.12
future~=1.0.0
gdown>=5.1.0,<6.0.0
tensorboard~=2.18.0
thop~=0.1.1-2209072238
keras~=3.8.0
inference_sdk
roboflow
timm

# =============================
# Face Recognition & Computer Vision
# =============================
face_recognition
opencv-python~=4.11.0.86
insightface==0.7.3
onnxruntime~=1.18.0

# =============================
# Tracking, Detection, Analytics
# =============================
boxmot==13.0.17
supervision==0.25.1
deep_sort_realtime==1.3.2

# =============================
# FastAPI & Related
# =============================
fastapi==0.116.0
uvicorn==0.35.0
loguru==0.7.3
pydantic~=2.11.5

# =============================
# OCR & NLP
# =============================
pytesseract==0.3.13
easyocr==1.7.2
transformers==4.53.1

# =============================
# Qdrant & Cloudinary
# =============================
qdrant-client==1.14.3
cloudinary==1.44.1

# =============================
# Environment/Build Tools
# =============================
cmake

# =============================
# Other dependencies
# =============================
python-dateutil~=2.9.0.post0
six~=1.17.0
yt-dlp>=2024.1.0
jsonschema
tenacity~=8.0.0
